refactor: extract frontend components
